add disable_mdns feature
This will disable mdns by nullifying it's static pod definition.
mdns is not strictly needed in envs where all hosts have dns A record.
disabling it will reduce amount of multicasts in the network, which is important in envs with mid-large amount of nodes.

The disable_mdns is false by default, retaining original behavior